Update India generation data with more years
Update build_database_IND.py with more capturing of annual generation
from the CEA. Also fixed a mis-attribution of generation to 2016
that should have been 2015. With this change India data from CEA
is now captured for 3 years (2013-2015).
This also affects the generation estimation for other Indian plants.

Still need to handle India reporting on fiscal year, not calendar.
Calling that a feature not a bug for now.

Add output_database/DATABASE_VERSION to be explicit about the database
versioning and record this with the git repo.