Update AUS with 2016-2017 generation & AREMI source
Update to Australia primarily, using a newly-fetched AREMI
XML file (which has changed its structure since it was last updated).
Identified and removed some duplicate plants in the AREMI dataset,
which caused the overall capacity to decrease a share, despite some new
plants also being included in the dataset.

Add the 2016-2017 NGER generation dataset, with an accompanying
matching table in `resources/AUS`. Still some work to be done
on these matches, but this should provide a good foundation for
inclusion of additional NGER years. The reporting time for this report
is July -> July, so this poorly maps to a calendar year. Assigning it to
2017 because that makes the most sense for now.

Add a new field to the output database: `generation_gwh_2017`.
With this, there are a couple new generation values present in the
output database for countries that are not Australia.
Updated the scripts and functions to handle this new generation year.

Net change: +10 plants @~ -0.45GW

Database Release: v1.1.4