Update AUS with historical generation from NGER
Update Australia with NGER generation for previous years.
Still maintaining the convention of prescribing the generation to the
second half of the fiscal year.

No net change in number of plants or capacity.

Some changes in generation values from previous database versions, which
occur primarily due to greater time to inspect the NGER dataset and with
additional context from previous years.