Add WEPP Location ID to database (PLATTS ID)
Addressing part of Issue #6, add a new field (`wepp_id`) to the output
database. This field holds a reference to a unique plant identifier in
the widely-used PLATTS-WEPP dataset. This should allow for
intercomparisons with an independently derived dataset, and will be
helpful to users who want to build a composite dataset for their
analyses.

The `wepp_id` field is a delimited field, with LOCATIONID values
separated by the pipe character (|) if multiple LOCATIONIDs match to a
single plant.

With this update, there are 16460 plants with known `wepp_id`s, which
account for 4312.6 GW in total (about 80% of all capacity in the output
database).

There are no changes to plants in the database except for the addition
of this `wepp_id`.

Database Version: 1.1.10