Remove Google Fusion Table dependency.
Google Fusion Tables are deprecated and "turned down"
as of 2019-12-03 [0]. This commit removes references to Fusion Tables
throughout the codebase.

The new manually-collected-data strategy is to hold the same local
copies in the git repo and not employ online, editable, exportable data
tables.

Update the python requirements to use `python-dateutil` rather than
`dateutil` since it seems the name has been changed on PyPi (pip).

This commit does not modify any power plant data or result in any
changes in the database.