Add JRC-PPDB-OPEN for some EU generation.
Add a new data source: JRC-PPDB-OPEN
(https://ec.europa.eu/jrc/en/publication/joint-research-centre-power-plant-database-jrc-ppdb)

This source from the Joint Research Centre of the European Union links a
few datasets together, including the Global Power Plant Database with
ENTSO-E.

Since JRC-PPDB-OPEN includes a measurement of the coverage of reported
data (data gaps may exist in the ENTSO-E reporting for any generating
unit), a decision was made such that all generating units for a
particular GPPD plant would need to have 95% time coverage for a year to
have that information aggregated and used as the 'reported' generation
value for a GPPD plant. This means there are far fewer plants with
reported generation than plants that are matched in JRC-PPDB-OPEN. This
should be remedied in the future and is a patch for an underlying data
representation issue.

Since some of the WRI-collected plants (formerly the 'Fusion Tables')
have existing generation values, but the source of those generation
values is not well identified, a new column is added to the 'WRI
country databases' which seeks to track the source of the singular
generation field in the WRI databases. This field has been partially
filled out for the accessible sources and should be given attention in
the future.

In total, this database adds the net generation as follows:
2015: +257 more plants with generation
2016: +400 more plants with generation
2017: +457 more plants with generation

This update does not affect the number of plants or total capacity of
the database.

Database version 1.2.2