Add `generation_data_source` column.
Add a new column to the output database - `generation_data_source`,
which describes the source of data for the `generation_gwh_YYYY`
columns. This field is required for any plant that has a reported
annual generation value. In the full data processing workflow, if a
plant has generation data but this data is not tied to a source, none of
the reported generation will be included in the final database.

For this reason, this update removes 51 instances of annual generation
which are no longer in the database. It is hoped that these will return
sometime in the future once the data sources are identified.

This database update does not change the total plant count or total
capacity of the Global Power Plant Database.

Database version 1.2.3