Add Legal Entity Identifier to database (GLEIF LEI)
Addressing part of issue 24, add a new field (`lei_code`) to the output
database. This field holds a reference to a plant's unique LEI in the GLEIF
dataset. This should allow users to connect plant data to LEI ownership
information and ISINs (unique financial instruments).

The `lei_code` field is an alphanumeric string matched to power plants
using the `master_gleif_concordance.csv` dataset.

With this update, there are 8132 plants with known `lei_code`s.

There are no changes to the database except for the addition of `lei_code`.

Database Version: 1.2.3