Remove percentage ownership from owner field
Bug: The Denmark.csv dataset contains data that quantifies ownership
with a percentage, rather than just the name of the entity. For example,
data provided by Vattenfall prepends owner names with their percent
ownership (100% Vattenfall instead of Vattenfall). This can cause
issues with programs that expect owner data to be the entity name only.

This commit fixes this bug.
There are 8 plants affected, all in Denmark.

Net change of 0 plants @ 0 capacity.

Database Version: 1.2.3