Update version to 1.3.0, announce project status.
This may a final commit to this project in an official capacity.
The hope is that because of the permissive licensing of data and code
others may benefit into the future.

This last update brings the repository generally in line with the June
2021 update to our technical note
(https://www.wri.org/research/global-database-power-plants)

The workflow, which now spans multiple repositories is described in the
README.